spark-minio_hive_metastore/requirements.txt
pyspark==3.5.1
python-dotenv==1.0.1
boto3==1.34.42
botocore==1.34.42
s3fs==2024.5.0
pyarrow==16.1.0
jinja2==3.1.3
pyodbc==5.0.1